+++ tests/Jackalope/Test/TestCase.php
@@ -8,29 +8,27 @@ use Jackalope\TestCase as BaseTestCase;
 
 abstract class TestCase extends BaseTestCase
 {
-    /**
-     * @var Connection
-     */
-    protected $conn;
+    protected ?Connection $conn;
 
     protected function getConnection(): Connection
     {
-        if (null === $this->conn) {
-            // @TODO see https://github.com/jackalope/jackalope-doctrine-dbal/issues/48
-            global $dbConn;
-            $this->conn = $dbConn;
-
-            if (null === $this->conn) {
-                $this->conn = DriverManager::getConnection([
-                    'driver' => @$GLOBALS['phpcr.doctrine.dbal.driver'],
-                    'path' => @$GLOBALS['phpcr.doctrine.dbal.path'],
-                    'host' => @$GLOBALS['phpcr.doctrine.dbal.host'],
-                    'user' => @$GLOBALS['phpcr.doctrine.dbal.username'],
-                    'password' => @$GLOBALS['phpcr.doctrine.dbal.password'],
-                    'dbname' => @$GLOBALS['phpcr.doctrine.dbal.dbname'],
-                ]);
-            }
+        if (isset($this->conn)) {
+            return $this->conn;
         }
+        // see https://github.com/jackalope/jackalope-doctrine-dbal/issues/48
+        global $dbConn;
+        if ($this->conn = $dbConn) {
+            return $this->conn;
+        }
+
+        $this->conn = DriverManager::getConnection([
+            'driver' => @$GLOBALS['phpcr.doctrine.dbal.driver'],
+            'path' => @$GLOBALS['phpcr.doctrine.dbal.path'],
+            'host' => @$GLOBALS['phpcr.doctrine.dbal.host'],
+            'user' => @$GLOBALS['phpcr.doctrine.dbal.username'],
+            'password' => @$GLOBALS['phpcr.doctrine.dbal.password'],
+            'dbname' => @$GLOBALS['phpcr.doctrine.dbal.dbname'],
+        ]);
 
         return $this->conn;
     }