+++ tests/bootstrap.php
@@ -1,6 +1,7 @@
 <?php
 
 use Doctrine\DBAL\DriverManager;
+use Doctrine\DBAL\Platforms\SqlitePlatform;
 use Jackalope\Transport\DoctrineDBAL\RepositorySchema;
 
 /* Make sure we get ALL infos from php */
@@ -43,7 +44,7 @@ $dbConn = DriverManager::getConnection([
 
 /* Recreate database schema */
 if (!getenv('JACKALOPE_NO_TEST_DB_INIT')) {
-    $options = ['disable_fks' => $dbConn->getDatabasePlatform() instanceof \Doctrine\DBAL\Platforms\SqlitePlatform];
+    $options = ['disable_fk' => $dbConn->getDatabasePlatform() instanceof SqlitePlatform];
     $repositorySchema = new RepositorySchema($options, $dbConn);
     $repositorySchema->reset();
 }